# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "CursorPage_ReviewCandidateView_": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewCandidateView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "CursorPage[ReviewCandidateView]",
        "type": "object"
      },
...
      "GraphStateChangeView": {
        "properties": {
          "after_status": {
            "anyOf": [
              {
                "enum": [
                  "learning",
                  "proficient",
                  "expert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Status"
          },
          "after_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Version"
          },
          "before_status": {
            "anyOf": [
              {
                "enum": [
                  "learning",
                  "proficient",
                  "expert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Status"
          },
          "before_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Version"
          },
          "changed_at": {
            "format": "date-time",
            "title": "Changed At",
            "type": "string"
          },
          "node_id": {
            "pattern": "^n\\d{3,}$",
            "title": "Node Id",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Reason Codes",
            "type": "array"
          },
          "source_type": {
            "enum": [
              "user",
              "summary_memory",
              "system_recompute"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "before_status",
          "after_status",
          "before_version",
          "after_version",
          "source_type",
          "changed_at"
        ],
        "title": "GraphStateChangeView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryNotification": {
        "properties": {
          "aggregate_id": {
            "title": "Aggregate Id",
            "type": "string"
          },
          "aggregate_type": {
            "title": "Aggregate Type",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "notification_id": {
            "format": "uuid",
            "title": "Notification Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "notification_id",
          "event_type",
          "title",
          "body",
          "aggregate_type",
          "aggregate_id",
          "read_at",
          "created_at"
        ],
        "title": "MemoryNotification",
        "type": "object"
      },
      "MemoryNotificationPage": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryNotification"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "unread_count": {
            "minimum": 0.0,
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_more",
          "unread_count"
        ],
        "title": "MemoryNotificationPage",
        "type": "object"
      },
      "MemoryOperationResult": {
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicError"
              },
              {
                "type": "null"
              }
            ]
          },
          "graph_state_changes": {
            "items": {
              "$ref": "#/components/schemas/GraphStateChangeView"
            },
            "title": "Graph State Changes",
            "type": "array"
          },
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/MutationResult"
            },
            "title": "Mutations",
            "type": "array"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "operation_type": {
            "enum": [
              "conversation_evidence",
              "activity_evidence",
              "correct_memory",
              "forget_memory",
              "restore_memory",
              "override_learner_profile",
              "review_candidate",
              "set_graph_state",
              "project_summary_to_graph",
              "rebuild_index",
              "verify_checksums",
              "purge_tombstones",
              "cleanup_orphan_versions",
              "cleanup_checkpoints",
              "purge_account_memory"
            ],
            "title": "Operation Type",
            "type": "string"
          },
          "review_candidate_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Review Candidate Ids",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "retry_wait",
              "succeeded",
              "needs_review",
              "dead_letter",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "operation_id",
          "status",
          "operation_type",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryOperationResult",
        "type": "object"
      },
...
      "PublicError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "trace_id"
        ],
        "title": "PublicError",
        "type": "object"
      },
...
      "ReviewDecisionRequest": {
        "additionalProperties": false,
        "description": "POST /review-candidates/{candidate_id}/decision 的公开请求体。\n\n与 GraphStatePutRequest 同模式（§6.4）：kind/candidate_id 由 Gateway\n固定注入后构造内部 ReviewCandidateCommand；客户端传入一律 422。",
        "properties": {
          "corrected_content": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "learner": "#/components/schemas/LearnerReplacement",
                    "mastery": "#/components/schemas/MasteryReplacement"
                  },
                  "propertyName": "replacement_type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/LearnerReplacement"
                  },
                  {
                    "$ref": "#/components/schemas/MasteryReplacement"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Content"
          },
          "decision": {
            "enum": [
              "accept",
              "correct",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "resolution_target": {
            "anyOf": [
              {
                "enum": [
                  "merge_existing",
                  "create_new_topic"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Target"
          },
          "target_memory_id": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Memory Id"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ReviewDecisionRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MemoryManagerGraph API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/memory/notifications": {
      "get": {
        "operationId": "list_notifications_api_v1_memory_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryNotificationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Notifications",
        "tags": [
          "memory-notifications"
        ]
      }
    },
    "/api/v1/memory/notifications/{notification_id}/read": {
      "post": {
        "description": "幂等已读（§19.6）；他人通知一律 404（§18.4）。",
        "operationId": "mark_notification_read_api_v1_memory_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryNotification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Notification Read",
        "tags": [
          "memory-notifications"
        ]
      }
    },
    "/api/v1/memory/operations/{operation_id}": {
      "get": {
        "description": "查询 operation 状态（§19.3）；他人 operation 一律 404（§18.4 IDOR）。",
        "operationId": "get_operation_api_v1_memory_operations__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Operation",
        "tags": [
          "memory-operations"
        ]
      }
    },
    "/api/v1/memory/operations/{operation_id}/cancel": {
      "post": {
        "description": "取消规则（§11.6）：queued/retry_wait 立即取消；running 协作取消；\n\nneeds_review 可取消；succeeded/dead_letter/cancelled 返回 409。",
        "operationId": "cancel_operation_api_v1_memory_operations__operation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Operation",
        "tags": [
          "memory-operations"
        ]
      }
    },
    "/api/v1/memory/review-candidates": {
      "get": {
        "operationId": "list_review_candidates_api_v1_memory_review_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "accepted",
                    "corrected",
                    "rejected",
                    "expired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_ReviewCandidateView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Review Candidates",
        "tags": [
          "memory-review-candidates"
        ]
      }
    },
    "/api/v1/memory/review-candidates/{candidate_id}/decision": {
      "post": {
        "description": "候选审核 P0 命令（§19.2）；候选归属与非 pending 状态在 Gateway 预检。",
        "operationId": "decide_review_candidate_api_v1_memory_review_candidates__candidate_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Review Candidate",
        "tags": [
          "memory-review-candidates"
        ]
      }
    },
...
  }
}